+++ src/blocks/FilterBlock/__stories__/FilterBlock.stories.tsx
@@ -1,60 +1,64 @@
 import {Meta, StoryFn} from '@storybook/react';
 
-import {yfmTransform} from '../../../../.storybook/utils';
+import {blockTransform} from '../../../../.storybook/utils';
 import {PageConstructor} from '../../../containers/PageConstructor';
-import {
-    FilterBlockModel,
-    FilterBlockProps,
-    FilterItem,
-    FilterTag,
-    LayoutItemModel,
-    LayoutItemProps,
-} from '../../../models';
+import {FilterBlockModel} from '../../../models';
 import FilterBlock from '../FilterBlock';
 
 import data from './data.json';
 
 export default {
-    title: 'Blocks/Filter Block',
+    title: 'Blocks/FilterBlock',
     component: FilterBlock,
+    argTypes: {
+        allTag: {
+            table: {
+                defaultValue: {
+                    summary: 'false',
+                },
+            },
+        },
+        tagButtonSize: {
+            table: {
+                defaultValue: {
+                    summary: '"l"',
+                },
+            },
+        },
+        centered: {
+            table: {
+                defaultValue: {
+                    summary: 'false',
+                },
+            },
+        },
+        animated: {
+            table: {
+                defaultValue: {
+                    summary: 'true',
+                },
+            },
+        },
+        colSizes: {
+            table: {
+                defaultValue: {
+                    summary: '{all: 12, sm: 6, md: 4};',
+                },
+            },
+        },
+    },
 } as Meta;
 
-const createItemList: (
-    count: number,
-    shared: LayoutItemProps,
-    tagList: FilterTag[],
-) => FilterItem[] = (count, shared, tagList) =>
-    Array.from({length: count}, (_, index) => ({
-        tags: [tagList[index % tagList.length].id],
-        card: {
-            ...shared,
-            content: {
-                title: shared.content.title
-                    ? `${shared.content.title}&nbsp;${index + 1}`
-                    : `${index + 1}`,
-            },
-        } as LayoutItemModel,
-    }));
-
-const createArgs = (overrides: Partial<FilterBlockProps>) =>
-    ({
-        type: 'filter-block',
-        title: data.default.content.title,
-        description: yfmTransform(data.default.content.description),
-        tags: data.default.filters,
-        items: createItemList(6, data.default.card, data.default.filters),
-        ...overrides,
-    }) as FilterBlockProps;
-
 const DefaultTemplate: StoryFn<FilterBlockModel> = (args) => (
-    <PageConstructor content={{blocks: [args]}} />
+    <PageConstructor content={{blocks: [blockTransform(args)]}} />
 );
 
 export const Default = DefaultTemplate.bind({});
-Default.args = createArgs({allTag: false});
-
-export const WithDefaultAllTag = DefaultTemplate.bind({});
-WithDefaultAllTag.args = createArgs({allTag: true});
-
+export const WithAllTag = DefaultTemplate.bind({});
 export const WithCustomAllTag = DefaultTemplate.bind({});
-WithCustomAllTag.args = createArgs({allTag: 'Custom All Tag Label'});
+export const Centered = DefaultTemplate.bind({});
+
+Default.args = data.default as FilterBlockModel;
+WithAllTag.args = {...data.default, allTag: true} as FilterBlockModel;
+WithCustomAllTag.args = {...data.default, allTag: 'Custom All Tag Name'} as FilterBlockModel;
+Centered.args = {...data.default, centered: true} as FilterBlockModel;