+++ src/blocks/Form/__tests__/Form.visual.test.tsx
@@ -7,6 +7,7 @@ import {
     FormData,
     WithBackgroundColor,
     WithBackgroundImage,
+    WithCustomFormNode,
 } from './helpers';
 
 const DEFAULT_FORM_DELAY = 20 * 1000;
@@ -45,6 +46,12 @@ test.describe('Form', () => {
 
     test.skip('render stories <FormData>', async ({mount, expectScreenshot, delay}) => {
         await mount(<FormData />);
+        await delay(DEFAULT_FORM_DELAY);
+        await expectScreenshot({skipTheme: 'dark'});
+    });
+
+    test.skip('render stories <WithCustomFormNode>', async ({mount, expectScreenshot, delay}) => {
+        await mount(<WithCustomFormNode />);
         await delay(DEFAULT_FORM_DELAY);
         await expectScreenshot({skipTheme: 'dark'});
     });