+++ src/sub-blocks/BasicCard/schema.ts
@@ -40,6 +40,9 @@ export const BasicCard = {
                 type: 'string',
                 enum: ['content', 'footer'],
             },
+            /**
+             * @deprecated This property will be removed in future versions
+             */
             hoverBackgroundColor: {
                 type: 'string',
             },