+++ src/sub-blocks/Content/__tests__/Content.test.tsx
@@ -5,6 +5,7 @@ import {
     testContentWithButtons,
     testContentWithCentered,
     testContentWithColSize,
+    testContentWithLabels,
     testContentWithLinks,
     testContentWithList,
     testContentWithSize,
@@ -31,6 +32,12 @@ const contentData: ContentProps = {
             text: 'list text',
         },
     ],
+    labels: [
+        {
+            icon: '/mock.png',
+            text: 'label text',
+        },
+    ],
     qa: 'content',
 };
 
@@ -39,7 +46,7 @@ const colSizesArray: GridColumnSizesType[] = [
     {all: 5, lg: 4, md: 3, sm: 2},
 ];
 
-const qaAttributes = getQaAttrubutes(contentData.qa, ['link', 'list']);
+const qaAttributes = getQaAttrubutes(contentData.qa, ['link', 'list', 'labels']);
 
 describe('Content', () => {
     test('Render by default', async () => {
@@ -128,6 +135,13 @@ describe('Content', () => {
         testContentWithList({
             props: contentData,
             options: {qaId: qaAttributes.list},
+        });
+    });
+
+    test('Render with labels', async () => {
+        testContentWithLabels({
+            props: contentData,
+            options: {qaId: qaAttributes.labels},
         });
     });
 });